Tenancy implementation first version

+++ DjeResidenceAPI/Models/Entities/UserSubscription.cs
@@ -3,7 +3,7 @@
     public class UserSubscription
     {
         public int Id { get; set; }
-        public int LanlordId { get; set; }
+        public long LanlordId { get; set; }
         public virtual ApplicationUser Lanlord { get; set; } = null!;
         public int SubscriptionPlanId { get; set; } 
         public virtual SubscriptionPlan SubscriptionPlan { get; set; } = null!;
